fixed bug that shows only whole units of hours

+++ parking/src/main/java/com/example/parking/DTO/VehicleExitResponse.java
@@ -6,10 +6,10 @@ public class VehicleExitResponse {
     private String plateNumber;
     private LocalDateTime entryTime;
     private LocalDateTime exitTime;
-    private long durationHours;
+    private double durationHours;
     private double fee;
 
-    public VehicleExitResponse(String plateNumber, LocalDateTime entryTime, LocalDateTime exitTime, long durationHours,
+    public VehicleExitResponse(String plateNumber, LocalDateTime entryTime, LocalDateTime exitTime, double durationHours,
             double fee) {
         this.plateNumber = plateNumber;
         this.entryTime = entryTime;
@@ -35,10 +35,10 @@ public class VehicleExitResponse {
     public void setExitTime(LocalDateTime exitTime) {
         this.exitTime = exitTime;
     }
-    public long getDurationHours() {
+    public double getDurationHours() {
         return durationHours;
     }
-    public void setDurationHours(long durationHours) {
+    public void setDurationHours(double durationHours) {
         this.durationHours = durationHours;
     }
     public double getFee() {
